feat(format)!: 切换 LFCA 4 统一静态合同
BREAKING CHANGE: Road Editing、LFCA、LFSM、LFSD 与 Identity registry 分别切换为 3、4、3、3 与 revision 3，并移除旧格式读取与 schema。

Refs: #551

+++ crates/laneflow-runtime/tests/cutover_cross_wall_clock_evidence.rs
@@ -114,7 +114,7 @@ fn world_with_fleet() -> TrafficWorld {
                 VehicleProfileOrdinal::from_raw(0),
                 route,
                 0,
-                1_000 + 6_500 * u32::from(index),
+                1_000 + 6_500 * index,
                 0,
             ))
             .expect("vehicle");
